Calculator 클래스에 제네릭 설정
int로 입력되던 값들을 모두 double로 변경

+++ src/myCalculator/Arithmetic.java
@@ -3,28 +3,28 @@ package myCalculator;
 public enum Arithmetic {
     // 덧셈 기능
     ADD("+") {
-        public int apply(int a, int b) {
+        public double apply(double a, double b) {
             return a + b;
         }
     },
 
     // 뺄셈 기능
     SUB("-") {
-        public int apply(int a, int b) {
+        public double apply(double a, double b) {
             return a - b;
         }
     },
 
     // 곱셈 기능
     MUL("*") {
-        public int apply(int a, int b) {
+        public double apply(double a, double b) {
             return a * b;
         }
     },
 
     // 나눗셈 기능
     DIV("/") {
-        public int apply(int a, int b) {
+        public double apply(double a, double b) {
             return a / b;
         }
     };
@@ -34,5 +34,5 @@ String symbol;
         this.symbol = symbol;
     }
 
-public abstract int apply(int a, int b);
+public abstract double apply(double a, double b);
 }
